requirements.txt
altgraph==0.17
asgiref==3.2.10
attrs==20.2.0
Automat==20.2.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
Django-3.1.4-py3-none-any.whl
django_channels-0.7.0-py3-none-any.whl
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
lxml==4.6.2
msgpack-python==0.5.6
oauthlib==3.1.0
pefile==2019.4.18
Pillow==7.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
pywin32-ctypes==0.2.0
redis==2.10.6
requests==2.25.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
sqlparse==0.3.1
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0
zope.interface==5.1.2
